Tau eff prints
I was wrong...
The optically thick medium is the one responsible for the curved shape
of the S graph. The thin one is behind the linear part of it. This does
not affect the result since the name of the variables is in fact the
right one. (the *_thick variables are the "first" critical point, the
one on the right of the S shape)
I have also printed out, in the output datafiles, the respective tau_eff
for T and S and for both of the critical points.
Furthermore, the T and S arrays given when calling the curve subroutine
are NOT dimensioned. (Just a reminder)