maps: Fix plots
The log scale on the plot is not relevant (the grid is not in log scale,
maybe we should change this), and the data were not plotted correctly
(reverse T axis, the correct fix is probably to be done in plot_map.py).

Also left to be done : contour plots showing S curve on each graphs.