main: Rework the logic.
The previous one was a bit confuse and from code changes to code
changes, several bit were forgotten, so it was potentially broken.

Now, just check whether all points are in the lower left cadrant defined
by the first critical point, if so do implicit integration of S,
integrate T until back on the curve and then Mdot kick to move the S
curve, else do explicit integration of both S and T.

Other changes: use prefactor only for S (it has no meaning for T) and
split timestep subroutine, cap prefactor to 0.1 instead of 1e-2 (this
last point is just a try).

With the current parameters, this code leads to the known numerical
instability.