s_curve: add brent method
It reduces the time to compute s_curve from 11s to 5s on my machine.
There is a bug now: if we activate brent and only brent for the
sigma_t_thick, we get an error. If we compute first using a dichotomy
then brent, it works.

This means that dichotomy has some unknown side-effects (which is really
bad), but I couldn't figure out what.

A quickfix is just to use dichotomy for thick case, and brent for thin
case. It converges faster (~2x less iterations than dichotomy).

I've also checked they yield the same results.