Change wasunstable use.
Only do Mdot un-kick when trying to get back on the curve. This might
need to be tweaked later if that’s too late, but just when going
unstable is too early for sure.

Also, this needed to fix initial was_unstable value (in both start and
restart/load cases). Only caveat I still see is if we restart just at
the point it starts becoming unstable…